Solved small warnings

+++ src/NeuralNetwork.cpp
@@ -57,7 +57,7 @@ void NeuralNetwork::feedForward(const std::vector<double> *inputValues) {
 	}
 
 	for (size_t i = 1; i < layers.size(); i++) {
-		for (int j = 0; j < layers[i].size(); j++) {
+		for (size_t j = 0; j < layers[i].size(); j++) {
 			layers[i][j].feedForward();
 		}
 	}
